Refactor to allow reusage from other client-code
